Task popup appears when 'add task' is clicked on when viewing a project. Restructure, and rename files for better readability.
